requirements.txt
# RBNS pipeline
requests>=2.28.0
pyyaml>=6.0
pandas>=1.5.0
